+++ src/Preprocess/ManualRot.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "SDL2_rotozoom.h"
 
-SDL_Surface *SimpleRot(SDL_Surface* image, float angle);
+SDL_Surface *SimpleRot(SDL_Surface* image, double angle);
